+++ packages/article/__tests__/ios/article.ios.test.js
@@ -1,14 +1,5 @@
 import shared from "../shared";
 
-jest.mock("react-native", () => {
-  const reactNative = require.requireActual("react-native");
-  reactNative.Platform.OS = "ios";
-  jest
-    .spyOn(reactNative.Platform, "select")
-    .mockImplementation(obj => obj.ios || obj.default);
-  return reactNative;
-});
-
 describe("Article test on ios", () => {
   shared();
 });