+++ packages/image/__tests__/ios/modal-image.ios.test.js
@@ -1,14 +1,5 @@
 import tests from "../modal-image.native.test";
 
-jest.mock("react-native", () => {
-  const reactNative = require.requireActual("react-native");
-  reactNative.Platform.OS = "ios";
-  jest
-    .spyOn(reactNative.Platform, "select")
-    .mockImplementation(obj => obj.ios || obj.default);
-  return reactNative;
-});
-
 describe("ModalImage test on ios", () => {
   tests();
 });