+++ packages/pull-quote/__tests__/android/pull-quotes.test.js
@@ -1,14 +1,5 @@
 import shared from "../shared";
 
-jest.mock("react-native", () => {
-  const reactNative = require.requireActual("react-native");
-  reactNative.Platform.OS = "android";
-  jest
-    .spyOn(reactNative.Platform, "select")
-    .mockImplementation(obj => obj.android || obj.default);
-  return reactNative;
-});
-
 describe("PullQuotes test on android", () => {
   shared();
 });